@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
-import {UserService} from "../../services/user.service";
-import {LoginService} from "../../services/login.service";
+import {UserService} from '../../services/user.service';
+import {LoginService} from '../../services/login.service';
 
 
 @Component({
@@ -10,10 +10,9 @@ import {LoginService} from "../../services/login.service";
 })
 export class HeaderComponent {
   constructor(public userService: UserService, private loginService: LoginService) {
-    this.userService.doMagic();
 
   }
-  logout(){
+  logout() {
     this.loginService.logout();
   }
 
