@@ -1,7 +1,6 @@
 import {Component, OnInit, ViewContainerRef} from '@angular/core';
 import {FormControl, FormGroup} from "@angular/forms";
 import {ForgotService} from "./forgot.service";
-import { ToastsManager } from 'ng2-toastr/ng2-toastr';
 
 @Component({
   selector: 'app-forgot',
@@ -12,9 +11,8 @@ export class ForgotComponent implements OnInit {
 
   private forgotService;
 
-  constructor(forgotService: ForgotService,public toastr: ToastsManager, vcr: ViewContainerRef) {
+  constructor(forgotService: ForgotService, public vcr: ViewContainerRef) {
     this.forgotService = forgotService;
-    this.toastr.setRootViewContainerRef(vcr);
 
   }
 
@@ -25,8 +23,8 @@ export class ForgotComponent implements OnInit {
 
 
 
-  RetypeEmail(){
-    this.forgotService.forgotPassword(this.retypeEmailFrom.controls.email.value)
+  RetypeEmail() {
+    this.forgotService.forgotPassword(this.retypeEmailFrom.controls.email.value);
 
   }
 
