@@ -1,7 +1,6 @@
 import { Injectable } from '@angular/core';
 import {Router} from "@angular/router";
 import {HttpClient} from "@angular/common/http";
-import { ToastsManager } from 'ng2-toastr/ng2-toastr';
 import Config from "../../../app-config";
 
 
@@ -11,7 +10,7 @@ export class ForgotService {
   private testUrl = Config.nodeApi;
   private error:boolean = false;
 
-  constructor( private http: HttpClient, private push: ToastsManager, private router:Router) {}
+  constructor( private http: HttpClient, private router:Router) {}
 
   forgotPassword(email){
     this.http
@@ -20,13 +19,13 @@ export class ForgotService {
         // Successful responses call the first callback.
         data => {
 
-          this.push.success('Email was send');
+          // this.push.success('Email was send');
 
         },
         // Errors will call this callback instead:
         err => {
 
-          this.push.error('Something went wrong, check if you spelled your email correctly');
+          // this.push.error('Something went wrong, check if you spelled your email correctly');
           console.log('Something went wrong!', err);
         }
       );
