@@ -1,12 +1,11 @@
 import { Injectable } from '@angular/core';
 import {Router} from "@angular/router";
 import {HttpClient} from "@angular/common/http";
-import { ToastsManager } from 'ng2-toastr/ng2-toastr';
 import Config from "../../../app-config"
 @Injectable()
 export class NewPasswordService {
 
-  constructor( private http: HttpClient, private push: ToastsManager, private router:Router) {}
+  constructor( private http: HttpClient, private router:Router) {}
 
   private testUrl = Config.nodeApi;
 
@@ -18,14 +17,14 @@ export class NewPasswordService {
         data => {
 
 
-            this.push.success('Password Succesfully changed');
+            // this.push.success('Password Succesfully changed');
 
             this.router.navigate(['login']);
 
         },
         // Errors will call this callback instead:
         err => {
-          this.push.error('Something went wrong');
+          // this.push.error('Something went wrong');
         }
       );
 }
